requirements.txt
Flask
requests
pandas
google-cloud-bigquery
google-auth
